Synthesizer

Documentation

synthesize :: Type -> TcM SDoc

Run the synthesis from a type and render the resulting expression

The result is in TcM because Type Hole Plugins use this monad and can SDoc can be directly used in a raw hole fit

synth :: Type -> Synth (HsExpr GhcPs)

Synthesize an expression from a type