Synthesizer.Monad

Synth

newtype Synth a

Synthesis monad

Handles threading of state, propagation of reading state, backtracking, constraint solving, and friends

Constructors

Synth

Fields

unSynth :: LogicT (RWST (Depth, [RestrictTag], Gamma, Omega) () (Int, Delta, [Ct]) TcM) a

runSynth :: Int -> Type -> Synth a -> TcM [a]

Run a synthesis computation.

Receives the amount of results to observe, the type of the recursive call (small hack while I don't have the module info), and the synthesis computation

Rules

type Rule = String

A Rule is a string e.g. "-oR", but could also eventually be a custom datatype

rule :: Rule -> Type -> Synth a -> Synth a

Run computation where the current rule is being applied to the given type

This is used to track the depth of the proof, and for logging

(=>>) :: Type -> Type -> Type

From two types (a,b) create a new type (a => b) for usage with rule

Used for representing e.g. left focus => synth goal

Gamma

newtype Gamma

Unrestricted hypothesis (Γ)

Constructors

Gamma

Fields

gamma :: [Prop]

inGamma :: Prop -> Synth a -> Synth a

Run a computation with additional proposition in gamma

getGamma :: Synth [Prop]

Get the unrestricted context

Delta

newtype Delta

Linear hypothesis (not left asynchronous) (Δ)

Constructors

Delta

Fields

delta :: [Prop]

pushDelta :: Prop -> Synth ()

Add a proposition to the linear context

delDelta :: Prop -> Synth ()

Remove a proposition from the linear context

getDelta :: Synth [Prop]

Get the linear context

setDelta :: [Prop] -> Synth ()

Set the linear context

guardUsed :: SName -> Synth ()

This computation succeeds when

Omega

newtype Omega

Ordered (linear?) hypothesis (Ω)

Constructors

Omega

Fields

omega :: [Prop]

takeOmegaOr :: Synth a -> (Prop -> Synth a) -> Synth a

Take a proposition from the omega context: * If one exists, pass it to the 2nd argument (computation using proposition) * If none does, run the 1st argument (computation without proposition from omega)

In case a proposition is taken from omega, the computation run won't have it in the omega context anymore

inOmega :: Prop -> Synth a -> Synth a

Run a computation with an additional proposition in omega

extendOmega :: [Prop] -> Synth a -> Synth a

Extend a computation with additional propositions in omega

Restrictions

data RestrictTag

Restrictions to conduct synthesis

Constructors

ConstructTy Type
Restriction on Constructing Type

DeconstructTy Type
Restriction on Deonstructing Type

DecideLeftBangR Type Type
Restriction on "Deciding-Left-Bang" on X to synthesize Y

addRestriction :: RestrictTag -> Synth a -> Synth a

Run a computation with an additional restriction

guardRestricted :: RestrictTag -> Synth ()

Guard that a certain restriction exists

i.e. if the given restriction is already in place, the computation suceeds, otherwise, the given restriction isn't in place, and the computation fails with empty

guardNotRestricted :: RestrictTag -> Synth ()

Guard that a certain restriction hasn't been applied

i.e. if the given restriction doesn't exist, the computation suceeds, otherwise, the given restriction is already in place, and the computation fails with empty

Constraints (WIP)

getConstraints :: Synth [Ct]

setConstraints :: [Ct] -> Synth ()

newWantedWithLoc :: CtLoc -> PredType -> TcM CtEvidence

solveConstraintsWithEq :: Type -> Type -> Synth ()

Add a constraint to the list of constraints and solve it Failing if the constraints can't be solved with this addition

Names

fresh :: Synth SName

Generate a fresh name

Example: a <- fresh return (var a)

Utilities

guardWith :: String -> Bool -> Synth ()

Like guard but emit given string as failure message upon failure

forMFairConj :: MonadLogic m => [t] -> (t -> m a) -> m [a]

forM but instead of (>>=) use (>>-) for fair conjunction

See (>>-) in LogicT

liftTcM :: TcM a -> Synth a

Lift a TcM computation into a Synth one

t4 :: (a, b, c, d) -> d

Get 4th value of tuple of size 4

Other Types

type Depth = Int

type SName = OccNameStr

type Prop = (SName, Type)

Orphan instances

Show Type

Instance details

Methods

showsPrec :: Int -> Type -> ShowS

show :: Type -> String

showList :: [Type] -> ShowS

Show Prop

Instance details

Methods

showsPrec :: Int -> Prop -> ShowS

show :: Prop -> String

showList :: [Prop] -> ShowS

Eq Type

Instance details

Methods

(==) :: Type -> Type -> Bool

(/=) :: Type -> Type -> Bool